% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faz_desenho_amostral_antigo.R
\name{faz_desenho_amostral_antigo}
\alias{faz_desenho_amostral_antigo}
\title{Funcao para definir desenho amostral usando metodo antigo (linearizacao de Taylor).}
\usage{
faz_desenho_amostral_antigo(data)
}
\arguments{
\item{data}{objeto contendo os microdados baixados da pnad.}
}
\description{
Funcao para definir desenho amostral usando metodo antigo (linearizacao de Taylor).
}
