% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faz_desenho_amostral_bootstrap.R
\name{faz_desenho_amostral_bootstrap}
\alias{faz_desenho_amostral_bootstrap}
\title{Funcao para definir desenho amostral usando pesos replicados (bootstrap).}
\usage{
faz_desenho_amostral_bootstrap(data)
}
\arguments{
\item{data}{objeto contendo os microdados baixados da pnad.}
}
\description{
Funcao para definir desenho amostral usando pesos replicados (bootstrap).
}
