% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faz_tab_freq.R
\name{faz_tab_freq}
\alias{faz_tab_freq}
\title{Funcao para fazer crosstabs de frequencia a partir da amostra complexa da Pnad}
\usage{
faz_tab_freq(data, ...)
}
\arguments{
\item{data}{objeto contendo os microdados da pnad, ja com o desenho da amostra commplexa configurado.}

\item{...}{nomes das variaveis usadas para recorte dos dados. Cria ainda uma variava "periodo" uma variavel com a informacao do trimestre. Ex.: 20201 = 1o trimestre de 2020.}
}
\description{
Funcao para fazer crosstabs de frequencia a partir da amostra complexa da Pnad
}
