% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faz_tab_media.R
\name{faz_tab_media}
\alias{faz_tab_media}
\title{Funcao para fazer crosstabs tendo como conteudo uma media, a partir da amostra complexa da Pnad}
\usage{
faz_tab_media(data, var_media, ...)
}
\arguments{
\item{data}{objeto contendo os microdados da pnad, ja com o desenho da amostra commplexa configurado.}

\item{var_media}{nome da variavel da qual sera extraida a media (conteudo da tabela).}

\item{...}{nomes das variaveis usadas para recorte dos dados. Cria ainda uma variava "periodo" uma variavel com a informacao do trimestre. Ex.: 20201 = 1o trimestre de 2020.}
}
\description{
Funcao para fazer crosstabs tendo como conteudo uma media, a partir da amostra complexa da Pnad
}
