% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importaPnad.R
\name{importaPnad}
\alias{importaPnad}
\title{Funcao para importar dados de um trimestre da pnad continua}
\usage{
importaPnad(trimestre, ano, bootstrap = F, lista_var = F)
}
\arguments{
\item{trimestre}{trimestre da pnad a ser baixada.}

\item{ano}{ano da pnad a ser baixada.}

\item{bootstrap}{logica, opcional. Se TRUE, faz o desenho amostral usando os pesos replicados (bootstrap). Se FALSE, faz o desenho amostral usando o metodo antigo (linearizacao de Taylor).}

\item{lista_var}{objeto opcional contendo lista de variaveis a serem baixadas.}
}
\description{
Funcao para importar dados de um trimestre da pnad continua
}
