% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importaPnad_sem_desenho_amostral.R
\name{importaPnad_sem_desenho_amostral}
\alias{importaPnad_sem_desenho_amostral}
\title{Funcao para importar dados de um trimestre da pnad continua}
\usage{
importaPnad_sem_desenho_amostral(
  trimestre,
  ano,
  bootstrap = F,
  lista_var = NULL,
  sm = F
)
}
\arguments{
\item{trimestre}{trimestre da pnad a ser baixada.}

\item{ano}{ano da pnad a ser baixada.}

\item{bootstrap}{logica, opcional. Se TRUE, mantem os pesos replicados para usar no bootstrap. Se FALSE, descarta essas 200 variaveis.}

\item{lista_var}{objeto opcional contendo lista de variaveis a serem baixadas.}

\item{sm}{logica, opcional. Se TRUE, acrescenta duas colunas que baixa do site do DIEESE: salario_minimo e salario_minimo_necessario.}
}
\description{
Funcao para importar dados de um trimestre da pnad continua
}
