% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_1x1.R
\name{media_1x1}
\alias{media_1x1}
\title{Funcao para fazer um crosstab de uma media, de 1x1 - com uma variavel na coluna e outra na linha}
\usage{
media_1x1(data, var_media, linha, coluna, cv_max = 0.3)
}
\arguments{
\item{data}{objeto contendo os microdados da pnad, ja com o desenho da amostra commplexa configurado.}

\item{var_media}{nome da variavel cuja media deve aparecer no conteudo da tabela.}

\item{linha}{nome da variavel das linhas.}

\item{coluna}{nome da variavel das colunas.}

\item{cv_max}{valor entre 0 e 1 para o coeficiente de variacao maximo aceito nas estimativas}
}
\description{
Funcao para fazer um crosstab de uma media, de 1x1 - com uma variavel na coluna e outra na linha
}
