% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importaPnad.R
\name{importaPnad}
\alias{importaPnad}
\title{Funcao para importar dados de um trimestre da pnad continua}
\usage{
importaPnad(trimestre, ano, bootstrap = F, lista_var = NULL)
}
\arguments{
\item{trimestre}{trimestre da pnad a ser baixada.}

\item{ano}{ano da pnad a ser baixada.}

\item{bootstrap}{logica, opcional. Se TRUE, faz o desenho amostral usando os pesos replicados (bootstrap). Se FALSE, faz o desenho amostral usando o metodo antigo (linearizacao de Taylor).}

\item{lista_var}{objeto opcional contendo lista de variaveis a serem baixadas.}
}
\description{
Funcao para importar dados de um trimestre da pnad continua
}
\examples{

# NAO RODAR!


# instala PnadDieese

# devtools::install_github("gustavoppmonteiro/PnadDieese")
library(PnadDieese)

# precisa desses pacotes:
library(tidyverse)
library(srvyr)
library(survey)
library(PNADcIBGE)


# seleciona variaveis
variaveis <-  c("V1028", "Ano", "Trimestre", "UF", "V2007", "VD4002")

# importa dados
dados_2T22 <- PnadDieese::importaPnad(2, 2022, lista_var = variaveis)

# tabela com frequencia de ocupados por sexo
T_1 <- dados_2T22 \%>\%
      filter(VD4002==1) \%>\%
      PnadDieese::faz_tab_freq(., V2007)

T_1
}
